pyproject.toml
[project]
name = "lesson-12-2"
version = "0.1.0"
description = ""
authors = [
    { name = "Dmitry_Zildermanis", email = "Uran.Deid@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^26.1.0"
isort = "^7.0.0"
mypy = "^1.19.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pre-commit = "^4.5.1"

[tool.black]
line-length = 119
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \__pycache__
  | \.tox
  | \.venv
  | dist
)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
explicit_package_bases = true
mypy_path = "src"
